Rename compliance_test.go to harness_test.go, absorb testutil_test.go
The file contains shared test infrastructure (MockUpstreamProxy,
ProxyUnderTest, helpers), not compliance tests. Merge the small
testutil_test.go (stubTokenProvider, testPseudonym) into the same file.
