+++ harness_test.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"strings"
 	"sync"
+	"sync/atomic"
 	"testing"
 	"time"
 )
@@ -482,4 +483,37 @@ func TestComplianceHarnessSmokeTest(t *testing.T) {
 
 	// Headers that should NOT appear on the forwarded request.
 	assertHeaderAbsent(t, upstreamReq.Header, "X-Nonexistent")
+}
+
+// ---------------------------------------------------------------------------
+// Shared test utilities
+// ---------------------------------------------------------------------------
+
+// testPseudonym is a fixed Via pseudonym used across tests for deterministic
+// assertions on the Via header value.
+const testPseudonym = "spnego-proxy-test"
+
+// stubTokenProvider is a controllable TokenProvider for testing.
+type stubTokenProvider struct {
+	calls  atomic.Int64
+	err    error         // when non-nil, GetToken returns this error
+	token  string        // returned on success
+	delay  time.Duration // artificial latency before returning
+	closed atomic.Bool
+}
+
+func (s *stubTokenProvider) GetToken() (string, error) {
+	s.calls.Add(1)
+	if s.delay > 0 {
+		time.Sleep(s.delay)
+	}
+	if s.err != nil {
+		return "", s.err
+	}
+	return s.token, nil
+}
+
+func (s *stubTokenProvider) Close() error {
+	s.closed.Store(true)
+	return nil
 }
